Copy approach from https://github.com/rylnd/monorepo-builds/

+++ api/deploy.sh
@@ -1,10 +1,14 @@
-docker pull fakenewsdetector/api
-docker build --cache-from fakenewsdetector/api -t fakenewsdetector/api .
-docker run --rm fakenewsdetector/api cargo test
+if [ -e .project-metadata/.api-changed ]; then
+    cd api
 
-if [ "${CIRCLE_BRANCH}" == "master" ]; then
-    docker push fakenewsdetector/api
+    docker pull fakenewsdetector/api
+    docker build --cache-from fakenewsdetector/api -t fakenewsdetector/api .
+    docker run --rm fakenewsdetector/api cargo test
 
-    docker tag fakenewsdetector/api registry.heroku.com/fake-news-detector-api/web
-    docker push registry.heroku.com/fake-news-detector-api/web
+    if [ "${CIRCLE_BRANCH}" == "master" ]; then
+        docker push fakenewsdetector/api
+
+        docker tag fakenewsdetector/api registry.heroku.com/fake-news-detector-api/web
+        docker push registry.heroku.com/fake-news-detector-api/web
+    fi
 fi
